Fixed printf-like format warnings.

+++ src/init.c
@@ -255,8 +255,8 @@ _dispatch_logv_init(void *context DISPATCH_UNUSED)
 			struct timeval tv;
 			gettimeofday(&tv, NULL);
 			fprintf(dispatch_logfile, "=== log file opened for %s[%u] at "
-					"%ld.%06u ===\n", getprogname() ?: "", getpid(),
-					tv.tv_sec, tv.tv_usec);
+					"%ld.%06lu ===\n", getprogname() ?: "", getpid(),
+					tv.tv_sec, (unsigned long)tv.tv_usec);
 			fflush(dispatch_logfile);
 		}
 	}
